+++ src/main/java/org/synyx/urlaubsverwaltung/security/ldap/LdapPersonContextMapper.java
@@ -13,6 +13,7 @@ import org.springframework.util.Assert;
 import org.synyx.urlaubsverwaltung.person.Person;
 import org.synyx.urlaubsverwaltung.person.PersonService;
 import org.synyx.urlaubsverwaltung.person.Role;
+import org.synyx.urlaubsverwaltung.security.PersonSyncService;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -20,6 +21,7 @@ import java.util.Optional;
 
 import static java.lang.invoke.MethodHandles.lookup;
 import static org.slf4j.LoggerFactory.getLogger;
+import static org.synyx.urlaubsverwaltung.person.Role.INACTIVE;
 
 
 /**
@@ -30,63 +32,54 @@ public class LdapPersonContextMapper implements UserDetailsContextMapper {
     private static final Logger LOG = getLogger(lookup().lookupClass());
 
     private final PersonService personService;
-    private final LdapSyncService ldapSyncService;
+    private final PersonSyncService personSyncService;
     private final LdapUserMapper ldapUserMapper;
 
-    LdapPersonContextMapper(PersonService personService, LdapSyncService ldapSyncService,
-                                   LdapUserMapper ldapUserMapper) {
-
+    LdapPersonContextMapper(PersonService personService, PersonSyncService personSyncService, LdapUserMapper ldapUserMapper) {
         this.personService = personService;
-        this.ldapSyncService = ldapSyncService;
+        this.personSyncService = personSyncService;
         this.ldapUserMapper = ldapUserMapper;
     }
 
     @Override
-    public UserDetails mapUserFromContext(DirContextOperations ctx, String username,
-        Collection<? extends GrantedAuthority> authorities) {
+    public UserDetails mapUserFromContext(DirContextOperations ctx, String username, Collection<? extends GrantedAuthority> authorities) {
 
-        LdapUser ldapUser;
+        final LdapUser ldapUser;
 
         try {
             ldapUser = ldapUserMapper.mapFromContext(ctx);
         } catch (InvalidSecurityConfigurationException | UnsupportedMemberAffiliationException ex) {
-            LOG.info("User '{}' can not sign in!", username, ex);
             throw new BadCredentialsException("No authentication possible for user = " + username, ex);
         }
 
-        String login = ldapUser.getUsername();
+        final String login = ldapUser.getUsername();
+        final Optional<String> firstName = ldapUser.getFirstName();
+        final Optional<String> lastName = ldapUser.getLastName();
+        final Optional<String> email = ldapUser.getEmail();
 
-        Optional<Person> optionalPerson = personService.getPersonByLogin(login);
+        final Person person;
 
-        Person person;
+        final Optional<Person> maybePerson = personService.getPersonByLogin(login);
+        if (maybePerson.isPresent()) {
+            final Person existentPerson = maybePerson.get();
 
-        if (optionalPerson.isPresent()) {
-            Person existentPerson = optionalPerson.get();
-
-            if (existentPerson.hasRole(Role.INACTIVE)) {
+            if (existentPerson.hasRole(INACTIVE)) {
                 LOG.info("User '{}' has been deactivated and can not sign in therefore", username);
 
                 throw new DisabledException("User '" + username + "' has been deactivated");
             }
 
-            person = ldapSyncService.syncPerson(existentPerson, ldapUser.getFirstName(), ldapUser.getLastName(),
-                    ldapUser.getEmail());
+            person = personSyncService.syncPerson(existentPerson, firstName, lastName, email);
         } else {
             LOG.info("No user found for username '{}'", username);
 
-            person = ldapSyncService.createPerson(login, ldapUser.getFirstName(), ldapUser.getLastName(),
-                    ldapUser.getEmail());
+            final Person createdPerson = personSyncService.createPerson(login, firstName, lastName, email);
+            person = personSyncService.appointAsOfficeUserIfNoOfficeUserPresent(createdPerson);
         }
 
         /*
          * NOTE: If the system has no office user yet, grant office permissions to successfully signed in user
          */
-        boolean noOfficeUserYet = personService.getPersonsByRole(Role.OFFICE).isEmpty();
-
-        // TODO: Think about if this logic could be dangerous?!
-        if (noOfficeUserYet) {
-            ldapSyncService.appointPersonAsOfficeUser(person);
-        }
 
         final Essence user = new Essence(ctx);
         user.setUsername(login);